requirements.txt
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.6.15.1
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.1
daphne==4.0.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.15
django-allauth==0.51.0
django-bootstrap5==22.1
django-debug-toolbar==3.7.0
django-environ==0.9.0
django-filter==22.1
django-location-field==2.1.0
django-pydenticon==0.2
django-redis==5.2.0
django-session-timeout==0.1.0
django-summernote==0.8.20.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
easy-thumbnails==2.8.3
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.4
oauthlib==3.2.2
packaging==21.3
Pillow==9.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydenticon==0.3.1
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
service-identity==21.1.0
simplejwt==2.0.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
Twisted==22.8.0
txaio==22.2.1
typing==3.7.4.3
typing_extensions==4.3.0
tzdata==2022.2
urllib3==1.26.12
webencodings==0.5.1
wincertstore==0.2
wrapt==1.14.1
zope.interface==5.4.0
django-bootstrap4
django-multiselectfield
django-taggit
django-extensions